+++ example/assignment_to_method_call.dart
@@ -1,4 +1,5 @@
 import 'package:mutator/mutator.dart';
+import 'dart:async';
 String code = """
     import 'package:mistletoe';
     var d = new Dynamism(expert:true);
@@ -12,7 +13,7 @@ const String pattern =
     '^[a-z.A-Z_0-9]+\\.on\\'
     '([a-z.A-Z_0-9]+\\)\\.[a-z.A-Z_0-9]+';
 String file_path = '';
-main() {
+main() async{
   replacer(e) {
     String s = e.toString();
     List l = s.split('=');
@@ -22,7 +23,7 @@ main() {
         '.set(\'${name}\', ${l.join('=').trim()})';
     return invocation;
   }
-  var m = new Mutator<AssignmentExpression>(
+  var m = await new Mutator<AssignmentExpression>(
       klass_name, pattern, replacer);
-  print(m.mutate_t(file_path, code: code));
+  print(await m.mutate_t(file_path, code: code));
 }